refactor(telemetry): extract computePercentiles into shared module
